New feature: Article merging
When there are multiple bloggers on one site, the authors may write more than
one article on the same topic and say similar things.  The user will be able to
visit the article view, and, if they are an admin, there should be a form on
that article’s edit page that allows an administrator to enter in the ID of
another article to merge the current article with.  When this happens, we want
to preserve both articles’ content, but merge them together into one article.
You should verify that both specified articles exist, and that they are not the
same article.

Scenarios
----------
* A non-admin cannot merge articles.
* When articles are merged, the merged article should contain the text of both
  previous articles.
* When articles are merged, the merged article should have one author (either
  one of the authors of the original article).
* Comments on each of the two original articles need to all carry over and
  point to the new, merged article.
* The title of the new article should be the title from either one of the
  merged articles.
* When articles are merged, the 2 merged articles are deleted.