Fixes a bug while accessing Categories in admin interface
The version of typo you have checked out contains a bug that prevents categories
from being created or edited. To see this bug, make sure you are logged in as an
administrator and click on the “categories” link on the left hand side of the
admin page.

A cucumber test is added that tests for this bug.